rust_protocol_basics

Trait ToHash32

pub trait ToHash32 {
    // Required method
    fn to_hash32(&self) -> Hash32;
}

A trait for types that can be converted to Hash32

This is useful for creating hashes from various data types.

Required Methods

fn to_hash32(&self) -> Hash32

Convert the type to a Hash32

Returns

A Hash32 representation of the data